% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{afghanistan_zone}
\alias{afghanistan_zone}
\title{Afghanistan}
\format{
A \verb{Simple feature collection} with 398 features and 4 fields:
\describe{
\item{country}{Country name}
\item{zone}{Zone name (health district)}
\item{area}{Zone area (m2)}
\item{zone_display}{Title case version of \code{zone}}
}
}
\source{
\href{https://www.geoboundaries.org}{geoBoundaries} via
\href{https://data.humdata.org/}{humdata} (2020).
}
\usage{
afghanistan_zone
}
\description{
Health district level sf data for Afghanistan.
}
\examples{
ggplot2::ggplot(afghanistan_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
