% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{drc_nat}
\alias{drc_nat}
\title{Democratic Republic of the Congo}
\format{
A \verb{Simple feature collection} with 1 feature and 3 fields:
\describe{
\item{country}{Country name}
\item{area}{Zone area (m2)}
\item{country_display}{Full and formatted name of \code{country}}
}
}
\source{
Legacy (ie. I don't know).
}
\usage{
drc_nat
}
\description{
National level sf data for the Democratic Republic of the Congo.
}
\examples{
ggplot2::ggplot(drc_nat) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
