% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{drc_reg}
\alias{drc_reg}
\title{DRC : Zone Level}
\format{
A \verb{Simple feature collection} with 519 features and 4 fields:
\describe{
\item{country}{Country name}
\item{reg}{Region (province) name}
\item{area}{Region area (m2)}
\item{reg_display}{Title case version of \code{reg}}
}
}
\source{
...unclear.
}
\usage{
drc_reg
}
\description{
Health zone level sf data for the Democratic Republic of the Congo.
}
\examples{
plot(sf::st_as_sf(drc_reg))
}
\keyword{dataset}
