% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{drc_zone}
\alias{drc_zone}
\title{DRC : Zone Level}
\format{
A \verb{Simple feature collection} with 519 features and 6 fields:
\describe{
\item{country}{Country name}
\item{reg}{Region (province) name}
\item{zone}{Zone name (health zone)}
\item{area}{Zone area (m2)}
\item{reg_display}{Title case version of \code{reg}}
\item{zone_display}{Title case version of \code{zone}}
}
}
\source{
...unclear.
}
\usage{
drc_zone
}
\description{
Health zone level sf data for the Democratic Republic of the Congo.
}
\examples{
plot(sf::st_as_sf(drc_zone))
}
\keyword{dataset}
