% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{liberia_reg}
\alias{liberia_reg}
\title{Liberia}
\format{
A \verb{Simple feature collection} with 15 features and 4 fields:
\describe{
\item{country}{Country name}
\item{reg}{Region (county) name}
\item{area}{Region area (m2)}
\item{reg_display}{Title case version of \code{reg}}
}
}
\source{
OCHA West and Central AFrica via \href{https://data.humdata.org/}{humdata} (2022).
}
\usage{
liberia_reg
}
\description{
Health zone level sf data for the Democratic Republic of the Congo.
}
\examples{
ggplot2::ggplot(liberia_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
