% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{liberia_zone}
\alias{liberia_zone}
\title{Liberia}
\format{
A \verb{Simple feature collection} with 136 features and 6 fields:
\describe{
\item{country}{Country name}
\item{reg}{Region (county) name}
\item{zone}{Zone name (health district)}
\item{area}{Zone area (m2)}
\item{reg_display}{Title case version of \code{reg}}
\item{zone_display}{Title case version of \code{zone}}
}
}
\source{
OCHA West and Central AFrica via \href{https://data.humdata.org/}{humdata} (2022).
}
\usage{
liberia_zone
}
\description{
Health zone level sf data for the Democratic Republic of the Congo.
}
\examples{
ggplot2::ggplot(liberia_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
