% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_map.R
\name{load_map}
\alias{load_map}
\title{Load map}
\usage{
load_map(
  country,
  level = c("zone", "reg", "nat"),
  as = c("sf", "tibble", "data.table")
)
}
\arguments{
\item{country}{\code{str} Country to load.}

\item{level}{\code{str} Level of geographic resolution to use, either : zone, reg (regional), nat
(national). Default is \code{'zone'}.}

\item{as}{\code{str} Type of output desired, either : \verb{'sf}', \code{'tibble'}, or \code{'data.table'}. Default
is \code{'sf'}.}
}
\description{
Returns a user-specified map in a either sf or tibble format.
}
\examples{
drc <- load_map('drc')
plot(drc)

drc <- load_map(country = 'drc',
                level = 'reg',
                as = 'data.table')
head(drc)

}
