% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{niger_old_zone}
\alias{niger_old_zone}
\title{Niger (Pre-2017 Districts)}
\format{
A \verb{Simple feature collection} with 36 features and 3 fields:
\describe{
\item{region}{Region Name}
\item{district}{District Name}
\item{pcode}{P Code}
}
}
\source{
https://geodata.lib.utexas.edu/catalog/stanford-bd373dw4168 (thanks Sophie!).
}
\usage{
niger_old_zone
}
\description{
Old (pre-2017) admin level 2 (district) level sf data for Niger.
}
\examples{
ggplot2::ggplot(niger_old_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
