% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{uganda_zone}
\alias{uganda_zone}
\title{Uganda}
\format{
A \verb{Simple feature collection} with 146 features and 6 fields:
\describe{
\item{adm0_name}{Country Name}
\item{adm0_iso3}{ISO Country Code}
\item{adm0_sub}{Country Subset (All, North, South, etc.)}
\item{adm1_name}{District Name}
\item{pcode}{P Code}
\item{adm1_pop}{District Population}
}
}
\source{
Epicentre DS.
}
\usage{
uganda_zone
}
\description{
Admin level 1 (district) level sf data for Uganda.
}
\examples{
ggplot2::ggplot(uganda_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
