% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{world}
\alias{world}
\title{The World}
\format{
A \verb{Simple feature collection} with 233 features and 2 fields
\describe{
\item{name}{Country name}
\item{continent}{Continent name}
}
}
\source{
\url{https://geojson-maps.ash.ms/}
}
\usage{
world
}
\description{
Countries of the World
}
\examples{
ggplot2::ggplot(world) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
