% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{drc_reg}
\alias{drc_reg}
\title{Democratic Republic of the Congo}
\format{
A \verb{Simple feature collection} with 26 features and 4 fields:
\describe{
\item{country}{Country name}
\item{reg}{Region (province) name}
\item{area}{Region area (m2)}
\item{reg_display}{Title case version of \code{reg}}
}
}
\source{
Legacy (ie. I don't know).
}
\usage{
drc_reg
}
\description{
Health zone level sf data for the Democratic Republic of the Congo.
}
\examples{
ggplot2::ggplot(drc_reg) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
