% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{liberia_nat}
\alias{liberia_nat}
\title{Liberia}
\format{
A \verb{Simple feature collection} with 1 feature and 3 fields:
\describe{
\item{country}{Country name}
\item{area}{Zone area (m2)}
\item{country_display}{Full and formatted name of \code{country}}
}
}
\source{
OCHA West and Central AFrica via \href{https://data.humdata.org/}{humdata} (2022).
}
\usage{
liberia_nat
}
\description{
National level sf data for Liberia.
}
\examples{
ggplot2::ggplot(liberia_zone) +
  ggplot2::geom_sf(fill = '#2e4473',
                   color = 'white') +
  ggplot2::theme_void()
}
\keyword{dataset}
